jacquard::sim::vcd_io

Function match_scope_path

pub fn match_scope_path<'i>(scope: &'i str, cur: &str) -> Option<&'i str>

Try to match one component in a scope path. Returns the remaining scope on success, or None on failure.